--- src/main/java/in/koreatech/koin/config/JpaConfiguration.java
@@ -1,10 +1,0 @@
-package in.koreatech.koin.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfiguration {
-
-}
