--- src/main/java/in/koreatech/koin/controller/GlobalExceptionHandler.java
@@ -1,25 +1,0 @@
-package in.koreatech.koin.controller;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler
-    public ResponseEntity<String> handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
-        log.warn(e.getMessage());
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
-        log.warn(e.getMessage());
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-}
