--- src/main/java/in/koreatech/koin/controller/UserController.java
@@ -1,26 +1,0 @@
-package in.koreatech.koin.controller;
-
-import in.koreatech.koin.dto.UserLoginRequest;
-import in.koreatech.koin.dto.UserLoginResponse;
-import in.koreatech.koin.service.UserService;
-import jakarta.validation.Valid;
-import java.net.URI;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserService userService;
-
-    @PostMapping("/user/login")
-    public ResponseEntity<UserLoginResponse> login(@RequestBody @Valid UserLoginRequest request) {
-        UserLoginResponse response = userService.login(request);
-        return ResponseEntity.created(URI.create("/"))
-            .body(response);
-    }
-}
