--- src/main/java/in/koreatech/koin/domain/TechStack.java
@@ -1,56 +1,0 @@
-package in.koreatech.koin.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Lob;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@Getter
-@Entity
-@Table(name = "tech_stacks")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class TechStack extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Lob
-    @Column(name = "image_url")
-    private String imageUrl;
-
-    @Size(max = 50)
-    @NotNull
-    @Column(name = "name", nullable = false, length = 50)
-    private String name;
-
-    @Size(max = 100)
-    @Column(name = "description", length = 100)
-    private String description;
-
-    @Column(name = "track_id")
-    private Long trackId;
-
-    @NotNull
-    @Column(name = "is_deleted", nullable = false)
-    private Boolean isDeleted = false;
-
-    @Builder
-    private TechStack(String imageUrl, String name, String description, Long trackId, Boolean isDeleted) {
-        this.imageUrl = imageUrl;
-        this.name = name;
-        this.description = description;
-        this.trackId = trackId;
-        this.isDeleted = isDeleted;
-    }
-}
