--- src/main/java/in/koreatech/koin/domain/Track.java
@@ -1,44 +1,0 @@
-package in.koreatech.koin.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Table(name = "tracks")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Track extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Size(max = 50)
-    @NotNull
-    @Column(name = "name", nullable = false, length = 50)
-    private String name;
-
-    @NotNull
-    @Column(name = "headcount", nullable = false)
-    private Integer headcount;
-
-    @NotNull
-    @Column(name = "is_deleted", nullable = false)
-    private Boolean isDeleted = false;
-
-    @Builder
-    private Track(String name) {
-        this.name = name;
-        this.headcount = 0;
-    }
-}
