+++ src/main/java/in/koreatech/koin/domain/track/repository/TechStackRepository.java
@@ -1,0 +1,14 @@
+package in.koreatech.koin.domain.track.repository;
+
+import java.util.List;
+
+import org.springframework.data.repository.Repository;
+
+import in.koreatech.koin.domain.track.domain.TechStack;
+
+public interface TechStackRepository extends Repository<TechStack, Long> {
+
+    TechStack save(TechStack techStack);
+
+    List<TechStack> findAllByTrackId(Long id);
+}
