--- src/main/java/in/koreatech/koin/domain/user/UserGender.java
@@ -1,7 +1,0 @@
-package in.koreatech.koin.domain.user;
-
-public enum UserGender {
-    MAN,
-    WOMAN,
-    ;
-}
