--- src/main/java/in/koreatech/koin/domain/user/UserToken.java
@@ -1,32 +1,0 @@
-package in.koreatech.koin.domain.user;
-
-import java.util.concurrent.TimeUnit;
-import lombok.Getter;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.redis.core.RedisHash;
-import org.springframework.data.redis.core.TimeToLive;
-
-@Getter
-@RedisHash("refreshToken")
-public class UserToken {
-
-    private static final long REFRESH_TOKEN_EXPIRE_DAY = 14L;
-
-    @Id
-    private Long id;
-
-    private final String refreshToken;
-
-    @TimeToLive(unit = TimeUnit.DAYS)
-    private final Long expiration;
-
-    private UserToken(Long id, String refreshToken, Long expiration) {
-        this.id = id;
-        this.refreshToken = refreshToken;
-        this.expiration = expiration;
-    }
-
-    public static UserToken create(Long userId, String refreshToken) {
-        return new UserToken(userId, refreshToken, REFRESH_TOKEN_EXPIRE_DAY);
-    }
-}
