+++ src/main/java/in/koreatech/koin/domain/user/domain/UserGender.java
@@ -1,0 +1,7 @@
+package in.koreatech.koin.domain.user.domain;
+
+public enum UserGender {
+    MAN,
+    WOMAN,
+    ;
+}
