+++ src/main/java/in/koreatech/koin/domain/user/domain/UserType.java
@@ -1,0 +1,18 @@
+package in.koreatech.koin.domain.user.domain;
+
+import lombok.Getter;
+
+@Getter
+public enum UserType {
+    STUDENT("STUDENT", "학생"),
+    USER("USER", "사용자"),
+    ;
+
+    private final String value;
+    private final String description;
+
+    UserType(String value, String description) {
+        this.value = value;
+        this.description = description;
+    }
+}
