--- src/main/java/in/koreatech/koin/dto/UserLoginRequest.java
@@ -1,16 +1,0 @@
-package in.koreatech.koin.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-
-@Getter
-public class UserLoginRequest {
-
-    @Email(message = "이메일 형식을 지켜주세요.")
-    @NotBlank(message = "이메일을 입력해주세요.")
-    private String email;
-
-    @NotBlank(message = "비밀번호를 입력해주세요.")
-    private String password;
-}
