+++ src/main/java/in/koreatech/koin/global/config/JpaConfiguration.java
@@ -1,0 +1,10 @@
+package in.koreatech.koin.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfiguration {
+
+}
